game-store-cloud/requirements.txt
Flask==3.0.0
google-cloud-firestore==2.14.0
cloud-sql-python-connector==1.11.0
pg8000==1.30.3
python-dotenv==1.0.0
pytest==7.4.3
werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0